+++ pdf/lib/src/widgets/chart/grid_cartesian.dart
@@ -28,7 +28,7 @@ class CartesianGrid extends ChartGrid {
   CartesianGrid({
     required GridAxis xAxis,
     required GridAxis yAxis,
-  })   : _xAxis = xAxis..direction = Axis.horizontal,
+  })  : _xAxis = xAxis..direction = Axis.horizontal,
         _yAxis = yAxis..direction = Axis.vertical;
 
   final GridAxis _xAxis;