+++ pdf/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fix Text.softWrap behavior
 - Add TableOfContent Widget
+- Add LinearProgressIndicator
 
 ## 3.3.0